% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uomthesisdown.R
\docType{package}
\name{uomthesisdown}
\alias{uomthesisdown}
\alias{uomthesisdown-package}
\title{uomthesisdown: A package for creating PhD theses using R Markdown}
\description{
uomthesisdown: A package for creating PhD theses using R Markdown
}
\section{thesis_gitbook}{

Creates an R Markdown thesis template as a webpage (great for writing)
}

\section{thesis_pdf}{

Creates an R Markdown thesis template as a PDF (looks lovely)
}

\section{thesis_word}{

Creates an R Markdown thesis template as a Microsoft Word document (looks
awful)
}

\section{thesis_epub}{

Creates an R Markdown thesis template as an ebook (looks awful)
}

